+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1026,7 +1026,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K5BAANYFVP5NHXS0SR12XB4K"),
+  id: z.string().default("01K5DS7B50TSZVG673GHJ1DK5J"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() => ListAgentsConfiguration2$inboundSchema),
@@ -1053,7 +1053,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K5BAANYFVP5NHXS0SR12XB4K"),
+  id: z.string().default("01K5DS7B50TSZVG673GHJ1DK5J"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() => ListAgentsConfiguration2$outboundSchema),