+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -2235,7 +2235,7 @@ export function httpFromJSON(
 /** @internal */
 export const Eleven$inboundSchema: z.ZodType<Eleven, z.ZodTypeDef, unknown> = z
   .object({
-    _id: z.string().default("01K5BAANZ2VKY1KWPSHAPE9WV1"),
+    _id: z.string().default("01K5DS7B5TD01V1SFT51HDT1XG"),
     path: z.string(),
     key: z.string(),
     display_name: z.string(),
@@ -2274,7 +2274,7 @@ export const Eleven$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Eleven
 > = z.object({
-  id: z.string().default("01K5BAANZ2VKY1KWPSHAPE9WV1"),
+  id: z.string().default("01K5DS7B5TD01V1SFT51HDT1XG"),
   path: z.string(),
   key: z.string(),
   displayName: z.string(),