+++ packages/orq-rc/RELEASES.md
@@ -3689,3 +3689,13 @@ Based on:
 - [typescript v3.13.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.15 - packages/orq-rc
+
+## 2025-09-18 06:48:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.618.1 (2.702.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.16 - packages/orq-rc